+++ imgui/src/input_widget.rs
@@ -3,6 +3,7 @@ use std::marker::PhantomData;
 use std::os::raw::{c_int, c_void};
 use std::ptr;
 
+use crate::internal::DataTypeKind;
 use crate::sys;
 use crate::{ImStr, ImString, Ui};
 
@@ -171,6 +172,17 @@ macro_rules! impl_text_flags {
     };
 }
 
+macro_rules! impl_format_params {
+    ($InputType:ident) => {
+        /// Sets the display format using *a C-style printf string*
+        #[inline]
+        pub fn display_format(mut self, display_format: &'p ImStr) -> Self {
+            self.display_format = Some(display_format);
+            self
+        }
+    };
+}
+
 macro_rules! impl_step_params {
     ($InputType:ident, $Value:ty) => {
         #[inline]
@@ -327,6 +339,91 @@ impl<'ui, 'p> InputTextMultiline<'ui, 'p> {
 }
 
 #[must_use]
+pub struct InputScalar<'ui, 'p, T: DataTypeKind> {
+    label: &'p ImStr,
+    step: Option<T>,
+    step_fast: Option<T>,
+    flags: InputTextFlags,
+    display_format: Option<&'p ImStr>,
+    _phantom: PhantomData<&'ui Ui<'ui>>,
+}
+
+impl<'ui, 'p, T: DataTypeKind> InputScalar<'ui, 'p, T> {
+    pub fn new(_: &Ui<'ui>, label: &'p ImStr) -> Self {
+        InputScalar {
+            label,
+            step: None,
+            step_fast: None,
+            flags: InputTextFlags::empty(),
+            display_format: None,
+            _phantom: PhantomData,
+        }
+    }
+
+    #[inline]
+    pub fn step(mut self, value: T) -> Self {
+        self.step = Some(value);
+        self
+    }
+
+    #[inline]
+    pub fn step_fast(mut self, value: T) -> Self {
+        self.step_fast = Some(value);
+        self
+    }
+
+    pub fn build(self, value: &mut T) -> bool {
+        unsafe {
+            sys::igInputScalar(
+                self.label.as_ptr(),
+                T::KIND as i32,
+                value as *mut T as *mut c_void,
+                self.step
+                    .as_ref()
+                    .map(|step| step as *const T)
+                    .unwrap_or(ptr::null()) as *const c_void,
+                self.step_fast
+                    .as_ref()
+                    .map(|step| step as *const T)
+                    .unwrap_or(ptr::null()) as *const c_void,
+                self.display_format
+                    .map(ImStr::as_ptr)
+                    .unwrap_or(ptr::null()),
+                self.flags.bits() as i32,
+            )
+        }
+    }
+
+    /// Builds a horizontal array of multiple drag sliders attached to the given slice.
+    ///
+    /// Returns true if any slider value was changed.
+    pub fn build_array(self, values: &mut [T]) -> bool {
+        unsafe {
+            sys::igInputScalarN(
+                self.label.as_ptr(),
+                T::KIND as i32,
+                values.as_mut_ptr() as *mut c_void,
+                values.len() as i32,
+                self.step
+                    .as_ref()
+                    .map(|step| step as *const T)
+                    .unwrap_or(ptr::null()) as *const c_void,
+                self.step_fast
+                    .as_ref()
+                    .map(|step| step as *const T)
+                    .unwrap_or(ptr::null()) as *const c_void,
+                self.display_format
+                    .map(ImStr::as_ptr)
+                    .unwrap_or(ptr::null()),
+                self.flags.bits() as i32,
+            )
+        }
+    }
+
+    impl_format_params!(InputScalar);
+    impl_text_flags!(InputScalar);
+}
+#[must_use]
 pub struct InputInt<'ui, 'p> {
     label: &'p ImStr,
     value: &'p mut i32,
@@ -371,6 +468,7 @@ pub struct InputFloat<'ui, 'p> {
     step: f32,
     step_fast: f32,
     flags: InputTextFlags,
+    display_format: Option<&'p ImStr>,
     _phantom: PhantomData<&'ui Ui<'ui>>,
 }
 
@@ -382,6 +480,7 @@ impl<'ui, 'p> InputFloat<'ui, 'p> {
             step: 0.0,
             step_fast: 0.0,
             flags: InputTextFlags::empty(),
+            display_format: None,
             _phantom: PhantomData,
         }
     }
@@ -393,12 +492,15 @@ impl<'ui, 'p> InputFloat<'ui, 'p> {
                 self.value as *mut f32,
                 self.step,
                 self.step_fast,
-                b"%.3f\0".as_ptr() as *const _,
+                self.display_format
+                    .map(ImStr::as_ptr)
+                    .unwrap_or(ptr::null()),
                 self.flags.bits() as i32,
             )
         }
     }
 
+    impl_format_params!(InputFloat);
     impl_step_params!(InputFloat, f32);
     impl_text_flags!(InputFloat);
 }
@@ -410,6 +512,7 @@ macro_rules! impl_input_floatn {
             label: &'p ImStr,
             value: &'p mut [f32; $N],
             flags: InputTextFlags,
+            display_format: Option<&'p ImStr>,
             _phantom: PhantomData<&'ui Ui<'ui>>,
         }
 
@@ -419,6 +522,7 @@ macro_rules! impl_input_floatn {
                     label,
                     value,
                     flags: InputTextFlags::empty(),
+                    display_format: None,
                     _phantom: PhantomData,
                 }
             }
@@ -428,12 +532,15 @@ macro_rules! impl_input_floatn {
                     sys::$igInputFloatN(
                         self.label.as_ptr(),
                         self.value.as_mut_ptr(),
-                        b"%.3f\0".as_ptr() as *const _,
+                        self.display_format
+                            .map(ImStr::as_ptr)
+                            .unwrap_or(ptr::null()),
                         self.flags.bits() as i32,
                     )
                 }
             }
 
+            impl_format_params!($InputFloatN);
             impl_text_flags!($InputFloatN);
         }
     };
@@ -442,6 +549,50 @@ macro_rules! impl_input_floatn {
 impl_input_floatn!(InputFloat2, 2, igInputFloat2);
 impl_input_floatn!(InputFloat3, 3, igInputFloat3);
 impl_input_floatn!(InputFloat4, 4, igInputFloat4);
+
+#[must_use]
+pub struct InputDouble<'ui, 'p> {
+    label: &'p ImStr,
+    value: &'p mut f64,
+    step: f64,
+    step_fast: f64,
+    flags: InputTextFlags,
+    display_format: Option<&'p ImStr>,
+    _phantom: PhantomData<&'ui Ui<'ui>>,
+}
+
+impl<'ui, 'p> InputDouble<'ui, 'p> {
+    pub fn new(_: &Ui<'ui>, label: &'p ImStr, value: &'p mut f64) -> Self {
+        InputDouble {
+            label,
+            value,
+            step: 1.0,
+            step_fast: 100.0,
+            flags: InputTextFlags::empty(),
+            display_format: None,
+            _phantom: PhantomData,
+        }
+    }
+
+    pub fn build(self) -> bool {
+        unsafe {
+            sys::igInputDouble(
+                self.label.as_ptr(),
+                self.value as *mut f64,
+                self.step,
+                self.step_fast,
+                self.display_format
+                    .map(ImStr::as_ptr)
+                    .unwrap_or(ptr::null()),
+                self.flags.bits() as i32,
+            )
+        }
+    }
+
+    impl_format_params!(InputDouble);
+    impl_step_params!(InputDouble, f64);
+    impl_text_flags!(InputDouble);
+}
 
 macro_rules! impl_input_intn {
     ($InputIntN:ident, $N:expr, $igInputIntN:ident) => {