+++ imgui/src/lib.rs
@@ -20,8 +20,8 @@ pub use self::fonts::glyph_ranges::*;
 pub use self::input::keyboard::*;
 pub use self::input::mouse::*;
 pub use self::input_widget::{
-    InputFloat, InputFloat2, InputFloat3, InputFloat4, InputInt, InputInt2, InputInt3, InputInt4,
-    InputText, InputTextMultiline,
+    InputDouble, InputFloat, InputFloat2, InputFloat3, InputFloat4, InputInt, InputInt2, InputInt3,
+    InputInt4, InputScalar, InputText, InputTextMultiline,
 };
 pub use self::io::*;
 pub use self::layout::*;
@@ -48,7 +48,7 @@ pub use self::widget::tab::*;
 pub use self::widget::tree::*;
 pub use self::window::child_window::*;
 pub use self::window::*;
-use internal::RawCast;
+use internal::{DataTypeKind, RawCast};
 
 #[macro_use]
 mod string;
@@ -264,10 +264,11 @@ impl<'ui> Ui<'ui> {
     ) -> InputTextMultiline<'ui, 'p> {
         InputTextMultiline::new(self, label, buf, size)
     }
-    #[doc(alias = "InputFloat2")]
+    #[doc(alias = "InputFloat")]
     pub fn input_float<'p>(&self, label: &'p ImStr, value: &'p mut f32) -> InputFloat<'ui, 'p> {
         InputFloat::new(self, label, value)
     }
+    #[doc(alias = "InputFloat2")]
     pub fn input_float2<'p>(
         &self,
         label: &'p ImStr,
@@ -290,6 +291,14 @@ impl<'ui> Ui<'ui> {
         value: &'p mut [f32; 4],
     ) -> InputFloat4<'ui, 'p> {
         InputFloat4::new(self, label, value)
+    }
+    #[doc(alias = "InputDouble")]
+    pub fn input_double<'p>(&self, label: &'p ImStr, value: &'p mut f64) -> InputDouble<'ui, 'p> {
+        InputDouble::new(self, label, value)
+    }
+    #[doc(alias = "InputScalar")]
+    pub fn input_scalar<'p, T: DataTypeKind>(&self, label: &'p ImStr) -> InputScalar<'ui, 'p, T> {
+        InputScalar::new(self, label)
     }
     #[doc(alias = "InputInt")]
     pub fn input_int<'p>(&self, label: &'p ImStr, value: &'p mut i32) -> InputInt<'ui, 'p> {